+++ sys/src/9/zynq/screen.h
@@ -6,8 +6,8 @@ struct Cursorinfo {
 };
 
 /* devmouse.c */
-extern void mousetrack(int, int, int, int);
-extern void absmousetrack(int, int, int, int);
+extern void mousetrack(int, int, int, ulong);
+extern void absmousetrack(int, int, int, ulong);
 extern Point mousexy(void);
 
 extern void mouseaccelerate(int);
